logを詳細に

+++ app/resolve/Python/rs_resolve/tool/voice_dropper/voice_dropper.py
@@ -297,7 +297,7 @@ class MainWindow(QMainWindow):
         ):
             f: Path
             QApplication.processEvents()
-            self.add2log('処理中: %s' % f.name)
+            self.add2log('Start: %s' % f.name)
             current_frame = get_currentframe(timeline)
 
             # キャラクター設定
@@ -330,6 +330,8 @@ class MainWindow(QMainWindow):
 
             # make timeline
             self.add2log('TMP TL: Start')
+
+            self.add2log('TMP TL: Create Temporary File: Start')
             tmp_tl_name = '_tmp_VoiceDropper_' + f.stem
             fcp_timeline = fcp.Timeline(self.xml)
             fcp_timeline.set_name(tmp_tl_name)
@@ -338,18 +340,22 @@ class MainWindow(QMainWindow):
             fcp_timeline.set_height(int(timeline.GetSetting('timelineResolutionHeight')))
             fcp_timeline.set_dropframe(timeline.GetSetting('timelineDropFrameTimecode') == '1')
             self.temp_file.write_text(str(fcp_timeline), encoding='utf-8')
+            if not self.temp_file.is_file():
+                self.add2log('Temporary Fileの作成に失敗しました。')
+                continue
+            self.add2log('TMP TL: Create Temporary File: Done')
 
             for i in range(1, project.GetTimelineCount() + 1):
                 tl = project.GetTimelineByIndex(i)
-                print(tl.GetName())
                 if tl.GetName() == tmp_tl_name:
                     media_pool.DeleteTimelines([tl])
-
+            self.add2log('TMP TL: Import Temporary File: Start')
             tmp_timeline = media_pool.ImportTimelineFromFile(str(self.temp_file))
             if tmp_timeline is None:
                 self.add2log('作業用Timelineの作成に失敗しました。')
                 continue
             self.add2log(f'TMP TL: {tmp_timeline.GetName()}')
+            self.add2log('TMP TL: Import Temporary File: Done')
 
             # time out 設定
             step = 0.2
@@ -393,7 +399,7 @@ class MainWindow(QMainWindow):
                 media_pool.DeleteTimelines([tmp_timeline])
                 continue
 
-            self.add2log('TMP TL: audio clip')
+            self.add2log('TMP TL:Insert Audio Clip: Start')
             # 音声トラックの選択
             if audio_index != 1:
                 send_hotkey(['ctrl', 'alt', str(audio_index)])
@@ -417,9 +423,9 @@ class MainWindow(QMainWindow):
                 media_pool.DeleteTimelines([tmp_timeline])
                 continue
             duration = clip.GetDuration()
-            self.add2log('TMP TL: audio clip insert')
+            self.add2log('TMP TL: Insert Audio Clip: Done')
 
-            self.add2log('TMP TL: text clip')
+            self.add2log('TMP TL: Insert Text Clip: Start')
             # text+クリップの仮挿入
             media_pool.AppendToTimeline([text_template])
             send_hotkey(['ctrl', 'z'])
@@ -432,9 +438,9 @@ class MainWindow(QMainWindow):
                 'endFrame': duration - 1,  # 1フレーム短くする (start 0 end 0 で 尺は1フレーム)
                 'mediaType': 1,
             }])[0]
-            self.add2log('TMP TL: text clip insert')
+            self.add2log('TMP TL: Insert Text Clip: Done')
 
-            self.add2log('TMP TL: lua script')
+            self.add2log('TMP TL: Run Lua Script: Start')
             # text+用のテキストを読み込み
             txt_file = f.parent.joinpath(f.stem + '.txt')
             if not txt_file.is_file() and data.make_text:
@@ -459,8 +465,8 @@ class MainWindow(QMainWindow):
             if ch_data.color in config.COLOR_LIST:
                 text_plus.SetClipColor(ch_data.color)
                 clip.SetClipColor(ch_data.color)
-            self.add2log('TMP TL: lua script done')
-            self.add2log('TMP TL: copy and paste')
+            self.add2log('TMP TL: Run Lua Script: Done')
+            self.add2log('TMP TL: Copy and Paste: Start')
             # リンク、コピー
             send_hotkey(['ctrl', '4'])
             send_hotkey(['ctrl', 'a'])
@@ -471,15 +477,16 @@ class MainWindow(QMainWindow):
             send_hotkey(['ctrl', '4'])
             set_currentframe(timeline, current_frame)
             send_hotkey(['ctrl', 'v'])
-            self.add2log('TMP TL: copy and paste done')
+            self.add2log('TMP TL: Copy and Paste: Done')
 
             # 再生ヘッドの移動
             set_currentframe(timeline, current_frame + duration + data.offset)
 
             self.add2log('Import: ' + str(f))
             # 終了処理
+            self.add2log('TMP TL: Clean Up: Start')
             media_pool.DeleteTimelines([tmp_timeline])
-            self.add2log('TMP TL: delete')
+            self.add2log('TMP TL: Clean Up: Done')
             #
             self.add2log('')
         # end
